backend/requirements.txt
Flask==2.2.3
flask-cors==4.0.0
supabase==1.2.0
python-dotenv==1.0.0
requests==2.31.0
librosa
numpy
python-multipart
gunicorn